backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
PyJWT
email-validator
